minor bug fix prior to merge with master

+++ db/migrate/20170901132302_drop_views.rb
@@ -1,5 +1,5 @@
 class DropViews < ActiveRecord::Migration
   def change
-  	drop_table :views
+  	drop_table :views if ActiveRecord::Base.connection.table_exists? 'views'
   end
 end
